Fix workspace Clippy warnings

+++ majit/majit-metainterp/src/optimizeopt/intdiv.rs
@@ -17,7 +17,7 @@ use crate::optimizeopt::OptContext;
 pub fn magic_numbers(m: i64) -> (u64, u32) {
     debug_assert!(m >= 3);
     debug_assert!(m & (m - 1) != 0, "m must not be a power of two");
-    debug_assert!(m == m as i64, "m must fit in signed i64"); // always true
+    debug_assert!(m == m, "m must fit in signed i64"); // always true
 
     let m_u = m as u64;
