+++ pyre/pyre-object/src/bufferview.rs
@@ -63,7 +63,7 @@ fn copy_rec(
 unsafe fn read_dims(t: PyObjectRef) -> Vec<i64> {
     unsafe {
         let n = crate::tupleobject::w_tuple_len(t);
-        let mut dims: Vec<i64> = Vec::with_capacity(n as usize);
+        let mut dims: Vec<i64> = Vec::with_capacity(n);
         for i in 0..n {
             let d = crate::tupleobject::w_tuple_getitem(t, i as i64)
                 .map(|w| crate::intobject::w_int_get_value(w))
